+++ lib/HOC/with-analytics.ts
@@ -1,4 +1,4 @@
-import analyticsContextTypes from './analytics-context-types';
+import analyticsContextTypes from '../analytics-context-types';
 import { getContext, InferableComponentEnhancerWithProps } from 'recompose';
 
 export const withAnalytics = getContext(analyticsContextTypes);
