+++ lib/index.ts
@@ -1,0 +1,7 @@
+export {withAnalytics} from './HOC/with-analytics';
+export {withoutAnalytics} from './HOC/without-analytics';
+export {enrichAnalytics} from './HOC/enrich-analytics';
+export {withAnalyticOnEvent} from './HOC/with-analytic-on-event';
+export {withAnalyticOnMount} from './HOC/with-analytic-on-mount';
+export {withOnPropChangedAnalytic} from './HOC/with-on-prop-changed-analytic';
+export {default as analytics} from './analytics';
